fix(epoch-oracle): greatly reduce base layer epoch sync times
