+++ Cargo.toml
@@ -1,4 +1,3 @@
-# NOTE: When editing this version, also edit the versions in template_built_in/templates/account and account_nft
 [workspace.package]
 version = "0.17.0"
 edition = "2021"
@@ -177,6 +176,7 @@ ethnum = "1.5.0"
 either = "1.13.0"
 fern = "0.7.1"
 futures = "0.3.30"
+futures-util = "0.3"
 futures-bounded = "0.2.3"
 jfs = "0.7.1"
 json5 = "0.4.1"
