+++ applications/tari_ootle_app_utilities/src/epoch_oracle_config.rs
@@ -48,7 +48,10 @@ impl Default for BaseLayerOracleConfig {
     fn default() -> Self {
         Self {
             base_node_grpc_url: None,
-            scanning_interval: Duration::from_secs(5),
+            // NOTE: this essentially sets the maximum possible lag (assuming the base node is perfectly in sync) for
+            // epoch change. Therefore, the `epoch_end_grace_period` in consensus settings should always be
+            // greater than this value to avoid a race condition resulting in leader failure.
+            scanning_interval: Duration::from_secs(9),
         }
     }
 }
