+++ applications/tari_validator_node/src/bootstrap.rs
@@ -42,7 +42,7 @@ use tari_epoch_manager::{
     EpochManagerReader,
 };
 use tari_epoch_oracles::{
-    base_layer::BaseLayerOracle,
+    base_layer::{BaseLayerEpochOracleConfig, BaseLayerOracle},
     configured::{ConfiguredEpochOracle, RealTimeEpochTicker},
     hybrid::{watch_ticker, HybridEpochOracle},
     store::EpochOracleStore,
@@ -522,23 +522,17 @@ async fn create_base_layer_epoch_oracle<TStore: EpochOracleStore + 'static>(
     Ok(BaseLayerOracle::new(
         store,
         base_node_client,
-        consensus_constants.base_layer_confirmations,
-        config.epoch_oracle.base_layer.scanning_interval,
-        config
-            .validator_node
-            .validator_node_sidechain_id
-            .as_ref()
-            .map(|p| p.to_byte_type()),
-        config
-            .validator_node
-            .burnt_utxo_sidechain_id
-            .as_ref()
-            .map(|p| p.to_byte_type()),
-        config
-            .validator_node
-            .template_sidechain_id
-            .as_ref()
-            .map(|p| p.to_byte_type()),
+        BaseLayerEpochOracleConfig {
+            start_height: 0,
+            height_lag: consensus_constants.base_layer_confirmations,
+            scanning_interval: config.epoch_oracle.base_layer.scanning_interval,
+            sidechain_id: config
+                .validator_node
+                .validator_node_sidechain_id
+                .as_ref()
+                .map(|p| p.to_byte_type()),
+            sync_headers: true,
+        },
         config.network,
     ))
 }
