+++ clients/base_node_client/Cargo.toml
@@ -20,7 +20,9 @@ log = { workspace = true }
 serde = { workspace = true, default-features = true }
 thiserror = { workspace = true }
 tonic = { workspace = true }
+futures-util = { workspace = true }
 ts-rs = { workspace = true, optional = true }
+tokio = { workspace = true, features = ["sync"] }
 url = { workspace = true }
 
 [features]
