+++ clients/base_node_client/src/traits.rs
@@ -6,9 +6,10 @@ use std::future::Future;
 use minotari_app_grpc::tari_rpc::ValidatorNodeChange;
 use tari_common_types::types::FixedHash;
 use tari_node_components::blocks::BlockHeader;
-use tari_ootle_common_types::{Epoch, SubstateAddress};
+use tari_ootle_common_types::Epoch;
 use tari_template_lib::prelude::RistrettoPublicKeyBytes;
 use tari_transaction_components::transaction_components::CodeTemplateRegistration;
+use tonic::codegen::tokio_stream::Stream;
 
 use crate::{
     error::BaseNodeClientError,
@@ -27,12 +28,9 @@ pub trait BaseNodeClient: Send + Sync + Clone {
     fn get_validator_nodes(
         &mut self,
         height: u64,
-    ) -> impl Future<Output = Result<Vec<BaseLayerValidatorNode>, BaseNodeClientError>> + Send;
-    fn get_shard_key(
-        &mut self,
-        epoch: Epoch,
-        public_key: &RistrettoPublicKeyBytes,
-    ) -> impl Future<Output = Result<Option<SubstateAddress>, BaseNodeClientError>> + Send;
+    ) -> impl Future<
+        Output = Result<impl Stream<Item = Result<BaseLayerValidatorNode, BaseNodeClientError>>, BaseNodeClientError>,
+    > + Send;
     fn get_template_registrations(
         &mut self,
         start_hash: Option<FixedHash>,
@@ -42,6 +40,13 @@ pub trait BaseNodeClient: Send + Sync + Clone {
         &mut self,
         block_hash: &FixedHash,
     ) -> impl Future<Output = Result<BlockHeader, BaseNodeClientError>> + Send;
+    fn stream_headers(
+        &mut self,
+        from_height: u64,
+        limit: u64,
+    ) -> impl Future<
+        Output = Result<impl Stream<Item = Result<BlockHeader, BaseNodeClientError>> + Unpin, BaseNodeClientError>,
+    > + Send;
     fn get_consensus_constants(
         &mut self,
         tip: u64,
