+++ crates/epoch_oracles/src/store.rs
@@ -2,13 +2,19 @@
 //   SPDX-License-Identifier: BSD-3-Clause
 
 use serde::{de::DeserializeOwned, Serialize};
-use tari_ootle_common_types::NodeAddressable;
-use tari_ootle_storage::global::GlobalDb;
+use tari_common_types::types::FixedHash;
+use tari_node_components::blocks::BlockHeader;
+use tari_ootle_common_types::{Epoch, NodeAddressable};
+use tari_ootle_storage::global::{BlockHeaderModel, GlobalDb};
 use tari_ootle_storage_sqlite::global::SqliteGlobalDbAdapter;
 
 pub trait EpochOracleStore {
     fn get<T: DeserializeOwned>(&self, key: &[u8]) -> anyhow::Result<Option<T>>;
     fn set<T: Serialize>(&self, key: &[u8], value: &T) -> anyhow::Result<()>;
+    fn add_block_headers<I: IntoIterator<Item = (Epoch, FixedHash, BlockHeader)>>(
+        &self,
+        headers: I,
+    ) -> anyhow::Result<()>;
 }
 
 pub enum StoreKey {
@@ -16,7 +22,6 @@ pub enum StoreKey {
     BaseLayerLastScannedBlockHeight,
     BaseLayerLastScannedBlockHash,
     BaseLayerLastEpochHash,
-    BaseLayerNextBlockHash,
     ConfiguredIsInitialized,
     ConfiguredCurrentEpoch,
 }
@@ -28,7 +33,6 @@ impl StoreKey {
             Self::BaseLayerLastScannedBlockHash => b"base_layer.last_scanned_block_hash",
             Self::BaseLayerLastScannedBlockHeight => b"base_layer.last_scanned_block_height",
             Self::BaseLayerLastEpochHash => b"base_layer.last_epoch_hash",
-            Self::BaseLayerNextBlockHash => b"base_layer.next_block_hash",
             Self::ConfiguredIsInitialized => b"configured_oracle.is_initialized",
             Self::ConfiguredCurrentEpoch => b"configured_oracle.current_epoch",
         }
@@ -45,6 +49,25 @@ impl<TAddr: NodeAddressable> EpochOracleStore for GlobalDb<SqliteGlobalDbAdapter
     fn set<T: Serialize>(&self, key: &[u8], value: &T) -> anyhow::Result<()> {
         let mut tx = self.create_transaction()?;
         self.metadata(&mut tx).set_metadata(key, value)?;
+        tx.commit()?;
+        Ok(())
+    }
+
+    fn add_block_headers<I: IntoIterator<Item = (Epoch, FixedHash, BlockHeader)>>(
+        &self,
+        headers: I,
+    ) -> anyhow::Result<()> {
+        let mut tx = self.create_transaction()?;
+        let mut header_db = self.block_headers(&mut tx);
+        for (epoch, block_hash, header) in headers {
+            header_db.insert(BlockHeaderModel {
+                epoch,
+                height: header.height,
+                block_hash,
+                kernel_merkle_root: header.kernel_mr,
+                validator_node_merkle_root: header.validator_node_mr,
+            })?;
+        }
         tx.commit()?;
         Ok(())
     }
